// ecom-react-client/src/redux/action-types.ts
export enum ActionTypes {
  AUTH_SET_LOADIING = 'Auth:SET_LOADIING',
  AUTH_LOGIN_USER = 'Auth:LOGIN_USER',
  AUTH_LOGIN_ERROR = 'Auth:LOGIN_ERROR',
  AUTH_LOGOUT_USER = 'Auth:LOGOUT_USER',
  AUTH_ALTER_USER = 'Auth:ALTER_USER',
  /****/
  LAYOUT_TOGGLE_SIDEBAR = 'Layout:TOGGLE_SIDEBAR',
  LAYOUT_SET_SIDEBAR = 'Layout:SET_SIDEBAR',
  LAYOUT_TOGGLE_STORE_LOCATOR = 'Layout:TOGGLE_STORE_LOCATOR',
  LAYOUT_TOGGLE_SIDENAV = 'Layout:TOGGLE_SIDENAV',
}